# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MSC Backend API Documentation",
        "description": "Interactive Swagger API documentation for MSC Backend",
        "contact": {
            "name": "MSC Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
        "/dashboard/hr": {
            "get": {
                "tags": [
                    "HR Management"
                ],
                "summary": "Get HR Dashboard Stats",
                "operationId": "f3ce716d7a64a9775b25acd005a30ff2",
                "responses": {
                    "200": {
                        "description": "HR dashboard statistics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/manager/dashboard": {
            "get": {
                "tags": [
                    "Manager Management"
                ],
                "summary": "Get Manager Dashboard Stats",
                "operationId": "75a65c7ab9d480c82f28ee589aec04c3",
                "responses": {
                    "200": {
                        "description": "Manager dashboard statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "employees": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "active": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "inactive": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "tasks": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "pending": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "in_progress": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "completed": {
                                                            "type": "integer",
                                                            "example": 3
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employee/dashboard": {
            "get": {
                "tags": [
                    "Employee Management"
                ],
                "summary": "Get Employee Dashboard Stats",
                "operationId": "010454ac73c954b66a194fbb4749e2e5",
                "responses": {
                    "200": {
                        "description": "Employee dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Employee dashboard data retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "employee": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Employee Name"
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "employee1"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "employee@example.com"
                                                        },
                                                        "department_role": {
                                                            "type": "string",
                                                            "example": "Software Developer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "statistics": {
                                                    "properties": {
                                                        "total_tasks": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "pending_tasks": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "in_progress_tasks": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "completed_tasks": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "overdue_tasks": {
                                                            "type": "integer",
                                                            "example": 0
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter ONLY your personal access token (e.g., 1|xxxx...)",
                "scheme": "bearer"
            }
        }
    },
...
}